refactor: simplify StationDashboard by removing unused filtering logic and integrating SensorTable
- Removed complex filtering logic and state management from StationDashboard to streamline the component.
- Integrated SensorTable for improved sensor data display, enhancing user experience and code maintainability.
- Updated useList hook to ensure proper loading state management during sensor data fetching.
